async_compile: Fix the wait method to actually wait
This method never triggered. It's used in 2 tests and they pass, so no serious
concern.

Note that I did introduce and fix a latent bug, which is if we called
shutdown_compile_workers, jobs would crash with this change due to ready_future
being finished if we called wait.

However we only call wait in tests so that bug is fine.

The other behaviour, is that if you called shutdown, I believe we may
potentially block on your first triton compile after that, until the pool was
ready. This should correctly switch to direct mode, until the pool is ready on
later warmups.

ghstack-source-id: 2c9eb9049029110cb9f030ece247733b6e376f98
Pull Request resolved: https://github.com/pytorch/pytorch/pull/161561