[OpenReg] Add Develop Notes for Integrating New Backend into PyTorch
To facilitate the integration of the new backend, we plan to publish a new development note that details all the key components,hoping to speed up the development of other accelerators.

This PR is the beginning of this note, and involve the part of registration of operators and we will gradually improve it and keep in sync with OpenReg's code.

ghstack-source-id: 0e5d65d6db15d060422ab06205c2f31155d73dbc
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/158644