Add C++ function to accelerate DTensor.__new__
As the new comment explains, pybinding DispatchKeySet is expensive; one native function avoids the number of DispatchKeySets we incur pybind overhead for to 1.

Reviewers please advise on a better location for the new torch._C function; I just threw it in Module.cpp because I didn't know where to put it.

ghstack-source-id: 515e1c2a0e7fd4aa0211a5e695ed01f8ea2d43fa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/161588