[OpenReg] Migrate Accelerator Document from source/notes into source/accelerator
As the tile stated.

As the document grows, the content will become more and more, so in order to make it easier for users to read and easier for developers to maintain, we have split this file into several separate files and placed them in a dedicated directory called "accelerator".

ghstack-source-id: 0fa7e317d4ed2aa9fda74c5d48c2a27f92548452
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/161845