Add inline fast paths for SymInt operators
If SymInt::maybe_as_int() returns non-empty, then we get an inline
fast path. The philosophy here (as with the previous PR) is to
preserve performance in the "plain old ints" case.

Observed time spent in SymInt functions in computeStorageNBytes to
drop (and not cost shift elsewhere in the function) after this change,
profiling detach() using code similar to the benchmark from #160580
and Linux perf.

ghstack-source-id: a9cd4c1a108e3528af3ffc89a5728a7ba3b25b30
Pull Request resolved: https://github.com/pytorch/pytorch/pull/161586